+++ src/gwt/src/org/rstudio/studio/client/workbench/model/SessionInfo.java
@@ -271,6 +271,10 @@ public class SessionInfo extends JavaScriptObject
       return this.active_project_file;
    }-*/;
 
+   public final native String getActiveProjectName() /*-{
+      return this.active_project_name;
+   }-*/;
+
    public final FileSystemItem getActiveProjectDir()
    {
       String projFile = getActiveProjectFile();