+++ NEWS.md
@@ -6,6 +6,7 @@
 - Updated Electron to version 28.0.0 (#14055)
 - RStudio Desktop on Windows and Linux supports auto-hiding the menu bar. (#8932)
 - The GWT code can now be built with JDKs > 11 (#11242)
+- R projects can be given a custom display name in Project Options (#1909)
 
 #### Posit Workbench
 -
@@ -15,6 +16,7 @@
 - Fixed shortcut for inserting an assignment operator to work on non-US English keyboards (#12457)
 - Fixed an issue where the menubar would show on secondary windows if Alt key was pressed (#13973)
 - Fixed Windows installer to delete Start Menu shortcut during uninstall (#13936)
+- Fixed current Git branch not always showing correctly in external editor windows (#14029)
 - Fixed tooltip to show correct keyboard shortcut when hovering over URLs in the editor (#12504)
 - Fixed Save As dialog on Windows not showing Save As Type field when extensions are hidden (#12965)